requirements.txt
altgraph==0.17.4
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
packaging==24.1
PyYAML==6.0.2
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2